fix(api): fix category tree builder
I knew it was too simple to be correct. With the previous
implementation, in the case of a category structure such as Income ->
Paychecks / Side Gigs -> Photography, where Side Gigs happens to appear
before Paychecks, the function would terminate after descending to the
Photography category. Because the result set from the database would be
Income (Level 0), Side Gigs (Level 1), Photography (Level 2), Paychecks
(Level 1), after we've called GetEntryTree for Side Gigs, next is set to
Photography, as it's the next in the list. But Photography is level 2,
and Income is level 0, so the loop terminates, and we return Side Gigs
as the only child of Income.

By returning the tree entry as well as its index, after the call to
GetEntryTree for Side Gigs's children, we look for the next child of
Income at the index after the last child of Side Gigs (i.e. index 3,
where Paychecks is). This gets us the whole tree.